// src/features/users/types/user.ts
export interface User {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  name: string | null;
  company: string | null;
  blog: string;
  location: string | null;
  email: string | null;
  hireable: boolean | null;
  bio: string | null;
  twitter_username: string | null;
  public_repos: number;
  public_gists: number;
  followers: number;
  following: number;
  created_at: string;
  updated_at: string;
}

export interface SearchUsersResponse {
  items: User[];
  total_count: number;
  incomplete_results: boolean;
}

export interface UseUsersResult {
  users: User[];
  isLoading: boolean;
  error: Error | null;
  search: string;
  handleSearch: (term: string) => void;
}

export interface UseUserSearchOptions {
  enabled?: boolean;
}

export interface UseUserSearchParams {
  search: string;
  options?: UseUserSearchOptions;
}